requirements.txt
flask==2.3.2
flask-socketio==5.3.3
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.8.0
